requirements.txt
discord.py
crayons
requests
pymongo